vapi-wrapper/scripts/requirements.txt
psycopg2-binary>=2.9.6
python-dotenv>=0.20.0 